added book images
These images will be used in the books page, and book specific pages.

Since i have an image column for all the books, i can simply put the name of the image in set column, and fetch it from the images folder too be used in templates.

This might however kill some performance, so i'll see how i can optimize this in the future