new controller for books page
This controller uses the new repository method for fetching all the books.

I directly inject the repository in the method paramaters, and send them too the rendered twig template.

I noticed that i can't directly access the values of the object. I'll probably have to convert the object into an array in the future.