created my first fixture! :D
This fixtures adds data too the Books entity. It will serve as default dummy data, for when you need to quickly setup data in the books table.

I figured i'd add this, because earlier my database got corrupt, and i lost all the books dummy data.. This fixture will act as a back-up!