fixed not being able to show book description on pages
There was a paramater for the get function, and since i was trying to import the description info without giving a paramater, i kept getting an error.

I simply removed the paramater so that i can directly import the description of a book.

I think the makerbundle automatically added the paramater while creating the entity.