fixed account creation ps encryption not working [and improved controller dependency injection]
There were a couple of issues why i couldn't get the encryption working.

NOT IMPLENTING THE INTERFACE IN THE ENTITY:

Too make sure the passwords actually got encrypted, you need to implement the passwordAuthenticatedUserInterface into the entity that is used by the form.

If you have a getPassword in the entity, symfony will automatically know that's the property it needs to hash

USING INCORRECT DATA DURING ENCRYPTION PROCESS (sign up CONTROLLER):

I was supposed to fetch the password from the submitted form data and use it in the hashedPassword method. I for some reason tried to fetch it from the user object ($user) lol

SECURITY.YAML NOT BEING CORRECTLY CONFIGURED:

and finally, i had too add the users entity in the password_hasher section of the security.yaml file. Since this tells symfony how to hash passwords for a specific class.

[IMPROVED DEPENDENCY INJECTIONS signup page controller]:

In the controller, i noticed that i directly injected the dependencies into the controller methods, this can become pretty unmaintainable as the class grows bigger, hense why i decided too go for a constructor injection approach.

Since the injected dependencies are way more maintainable this way.