configured security.yaml file
Haven't been comitting as much since i'm taking some breaks during the summer break.

In order for the user to be logged in and have a user session, the security.yaml file needed to be configured to enable this.

I'm not too sure if it works yet so i'll have to do some testing to make sure it is configured properly